+++ include/algui_tree.h
@@ -54,7 +54,7 @@ ALGUI_TREE *algui_get_last_child_tree(ALGUI_TREE *tree);
     @param tree tree node to get the number of children.
     @return the number of children.
  */
-size_t algui_get_tree_child_count(ALGUI_TREE *tree);
+unsigned long algui_get_tree_child_count(ALGUI_TREE *tree);
 
 
 /** returns the data stored in a tree node.
